Sily changes

+++ js/common.js
@@ -1,13 +1,13 @@
 function getElementById(input) {
     const inputField = document.getElementById(input)
     const inputFieldString = inputField.value
-    const inputFieldValue = parseInt(inputFieldString)
+    const inputFieldValue = parseFloat(inputFieldString)
     return inputFieldValue
 }
 function getElementByTextId(input) {
     const inputField = document.getElementById(input)
     const inputFieldString = inputField.innerText
-    const inputFieldValue = parseInt(inputFieldString)
+    const inputFieldValue = parseFloat(inputFieldString)
     return inputFieldValue
 }
 function setTextOnTextField(input, value) {
